Create cybersecurity portfolio
This commit includes the initial version of my cybersecurity portfolio website. It features sections such as About Me, Experience, Projects, Certificates, Technical Toolkit, and a contact footer. The design is responsive and includes basic styling for improved aesthetics. The layout is structured with HTML and styled using CSS. Future updates will focus on enhancing the design and adding more projects and certifications.